+++ index.js
@@ -1,5 +1,12 @@
 /**
- * Takes any javascript value and returns a human readable string representation of that value. Designed for use in test messages, -0 is rendered as '-0', strings are wrapped in single quotes, and Arrays and Objects are passed through JSON.stringify.
+ *  Designed for use in test messages, displayValue takes a javascript value and returns a human readable string representation of that value.
+ *
+ * Notes:
+ * - -0 is rendered as -0
+ * - strings are wrapped in single quotes
+ * -  Arrays and Objects are passed through JSON.stringify
+ * - Array-like values such as arguments are handled like Arrays
+ * - Object-like values such as ClientRect and DOMRect are handled like Objects
  *
  * @example
  *
@@ -29,14 +36,17 @@
  *
  * @returns {string}
  */
-module.exports = (value, settings = {}) => {
+const displayValue = (value, settings = {}) => {
 	if (Object.is(value, -0)) {
 		return '-0';
 	}
 	if (value instanceof String || typeof value === 'string') {
 		return '\'' + value + '\'';
 	}
-	if (Array.isArray(value) || (value && value.constructor === Object)) {
+	if (value && (Array.isArray(value) || value.constructor === Object || (value.toJSON && !(value instanceof Date)))) {
+		if (typeof value.length === 'number') {
+			value = Array.prototype.slice.call(value);
+		}
 		if (settings.beautify) {
 			return JSON.stringify(value, null, 4);
 		}
@@ -45,3 +55,5 @@ module.exports = (value, settings = {}) => {
 
 	return value + '';
 };
+
+module.exports = displayValue;