+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [1.1.0] - 2019-09-20
+### Added
+- Supporting Array-like values
+- Supporting Object-like values
+
+## [1.0.2] - 2019-08-02
+### Security
+- Updating dependencies
+
+## [1.0.1] - 2019-07-07
+- documentation
+
 ## 1.0.0 - 2019-07-07
 ### Added
 - displayValue
+
+[1.1.0]: https://github.com/DarrenPaulWright/display-value/compare/v1.0.2...v1.1.0
+[1.0.2]: https://github.com/DarrenPaulWright/display-value/compare/v1.0.1...v1.0.2
+[1.0.1]: https://github.com/DarrenPaulWright/display-value/compare/v1.0.0...v1.0.1